+++ webknossos/webknossos/dataset/compress_utils.py
@@ -1,5 +1,5 @@
 import logging
-from os import makedirs, path
+import warnings
 from pathlib import Path
 from typing import Tuple
 
@@ -9,14 +9,15 @@ from ..utils import time_start, time_stop
 
 
 def compress_file_job(args: Tuple[Path, Path]) -> None:
+    warnings.warn("[DEPRECATION] compress_file_job is deprecated.", DeprecationWarning)
     source_path, target_path = args
     try:
         time_start("Compressing '{}' to '{}'".format(source_path, target_path))
 
-        makedirs(path.dirname(target_path), exist_ok=True)
+        target_path.parent.mkdir(parents=True, exist_ok=True)
         wkw.File.compress(str(source_path), str(target_path))
 
-        if not path.exists(target_path):
+        if not target_path.exists():
             raise Exception("Did not create compressed file {}".format(target_path))
 
         time_stop("Compressing '{}' to '{}'".format(source_path, target_path))