+++ webknossos/webknossos/utils.py
@@ -3,12 +3,14 @@ import calendar
 import functools
 import json
 import logging
-import os
+import sys
 import time
+import warnings
 from concurrent.futures import as_completed
 from concurrent.futures._base import Future
 from datetime import datetime
 from multiprocessing import cpu_count
+from os.path import relpath
 from pathlib import Path
 from typing import Any, Callable, Iterable, List, Optional, Union
 
@@ -71,6 +73,12 @@ def get_executor_for_args(
             job_resources=json.loads(args.job_resources),
         )
         logging.info(f"Using {args.distribution_strategy} cluster.")
+    elif args.distribution_strategy == "debug_sequential":
+        executor = get_executor(
+            args.distribution_strategy,
+            debug=True,
+            keep_logs=True,
+        )
     else:
         logging.error(
             "Unknown distribution strategy: {}".format(args.distribution_strategy)
@@ -140,17 +148,35 @@ def copy_directory_with_symlinks(
         if item.name not in ignore:
             symlink_path = dst_path / item.name
             if make_relative:
-                rel_or_abspath = Path(os.path.relpath(item, symlink_path.parent))
+                rel_or_abspath = Path(relpath(item, symlink_path.parent))
             else:
                 rel_or_abspath = item.resolve()
             symlink_path.symlink_to(rel_or_abspath)
 
 
 def setup_logging(args: argparse.Namespace) -> None:
-    logging.basicConfig(
-        level=(logging.DEBUG if args.verbose else logging.INFO),
-        format="%(asctime)s %(levelname)s %(message)s",
-    )
+    log_path = Path(f"./logs/cuber_{time.strftime('%Y-%m-%d_%H%M%S')}.txt")
+
+    console_log_level = logging.DEBUG if args.verbose else logging.INFO
+    file_log_level = logging.DEBUG
+
+    logging_formatter = logging.Formatter("%(asctime)s %(levelname)s %(message)s")
+
+    # Always set the global log level to the more verbose of console_log_level and
+    # file_log_level to allow to log with different log levels to console and files.
+    root_logger = logging.getLogger()
+    root_logger.setLevel(min(console_log_level, file_log_level))
+
+    console = logging.StreamHandler(sys.stdout)
+    console.setLevel(console_log_level)
+    console.setFormatter(logging_formatter)
+    root_logger.addHandler(console)
+
+    log_path.parent.mkdir(parents=True, exist_ok=True)
+    file_handler = logging.FileHandler(log_path, mode="w", encoding="UTF-8")
+    file_handler.setLevel(file_log_level)
+    file_handler.setFormatter(logging_formatter)
+    root_logger.addHandler(file_handler)
 
 
 def add_verbose_flag(parser: argparse.ArgumentParser) -> None:
@@ -169,4 +195,11 @@ def get_rich_progress() -> Progress:
         rich.progress.TimeElapsedColumn(),
         "|",
         rich.progress.TimeRemainingColumn(),
+    )
+
+
+def warn_deprecated(deprecated_item: str, alternative_item: str) -> None:
+    warnings.warn(
+        f"[DEPRECATION] `{deprecated_item}` is deprecated, please use `{alternative_item}` instead.",
+        DeprecationWarning,
     )